feat(Rendering): Add EdgeVisibility and Coincident support
Add code to support EdgeVisibility, EdgeColor,
CoincidentTopology options in the Property and Mapper
respectively. These values shoudl now be honored by
the OpenGLPolyDataMapper. At the same time consolidated
the primitives in the OpenGLPolyDataMapper to be an array
so that the code stay cleaner and easier to manage.

Minor change to the captureImage method to invoke it
directly from a view as the RenderWindow itself doesn't
really have a size or displayed convas/window associtated
with it.  Those are part of the view.

Addresses issue:

https://github.com/Kitware/vtk-js/issues/20