fix(Pipeline): fix an issue with shouldUpdate
The computation for when a filter should reexecute had
an issue that could cause it to always reexecute. This
change makes it so that a filter will always execute if
its inputs are more recent than its outputs or if its
mtime is more recent than its outputs. A filter is not
explicitly modified when executing and may have a mtime
that is lower than its inputs.