feat(Rendering): add hardware selector and framebuffers
This topic adds a hardwareslector class that can be used
to perform hardware picks and return the prop and propid.
In the future we can extend it to support the other passes
such as composite id and attribute id.

To support this we added a SelectionNode class that contains
information and properties about a selection.

Since webgl doesn't really have an option to turn of bufferswapping
we add a framebuffer class so that the hardware picking can take
place in the framebuffer instead of the visible buffer. The
framebuffer class will of course be useful for a lot of other
cases.

Fixed a bug where is a class initialized mtime to zero it
would not be honored and would instead use the global mtime.