feat(Rendering): Add support for volume rendering
This topic has a number of changes. The largest is that
it adds support for Volume rendering of single component
data with opacity and color transfer functions. The data
can be of any native type and it will be encoded into
multiple unsigned char bytes per value if needed to
preserve the resolution of the original data. If the
original volume is too large to fit in one texture then
it will be automatically subsampled in X and Y so that
it fits.

The volume rendering supports settnig both sample
distance and image sample distance so that you can
control rendering rates and performance. When
image sample distance is not unity a framebuffer
is used for rendering and then stretched into the
resulting window.

Volume rendering includes support for nearest
neightbor, fast linear, and linear interpolation.
Fast linear interpolates in X and Y but not Z.

Future feature may include lighting, opacity gradient
functions, intermixed opaque geometry, cropping planes,
and multicomponent input data.

This topic includes a fix for animations driven by
mouse movements. In the past mouse movements would
queue up rendering irregardless of the performance
of the graphcis card. This could cause sluggishness
and lurching behavior. The new code makes use of the
requestAnimationFrame functionality to let the browser
drive the rendering rate. Due to this change the
interactors will no longer respond to mouse move events
but rather animation events. Upon a mouse down they will
request aniation frames, respond to them, and upon mouse
up, cancel them.

A variety of minor bug fixes are included as well.