fix(Rendering): fix world to view conversions
The world to view conversions in Renderer were hardcoded to
a 1.0 aspect ratio and an incorrect zvalue range. This change
requires the aspect ratio to be passed in. As a convinience
when doing conversions that involve world -- view methods
have been added on the view (OpenGL/RenderWindow). These
methods take the renderer as an argument and include display -- world
and view -- world conversions.