fix(Rendering): fix imagemapper depth pass
The imagemapper was not handling the depth pass so
implement that. At the same time improve the polydatamapper
depth pass to not rebuild shaders every frame when
depth passes are requested. Instead it rebuilds the shader
the first time a depth pass is requested and then uses
a uniform from that point forward.