fix(macro): fix callback infinite loop
If new callbacks were added to the array during a callback (example: when
adding a widget that also listens to events), and if these callbacks have
a higher priority, the sorting of those callbacks would prepend them to the
array, which means the previous callback would be triggered again.

Using a copied array defined when `invoke` is called addresses that issue.