feat(RenderWindow): Properly handle image capture
replaced captureImage by captureNextImage which now return a promise on the OpenGL side.
On the Core/RenderWindow the method name remain the same but will return an array of promises.

BREAKING CHANGE: OpenGL/RenderWindow::captureNextImage returns a promise with the imageURL instead of returing it
directly. Core/RenderWindow::captureImages returns an array of promise.