feat(LookupTable): Support setTable to directly set colors
In order to support directly setting lookup table colors, support
`setTable`, which corresponds to the VTK C++ method:

  https://github.com/Kitware/VTK/blob/898d75800e5e33b6192b3f92b577715627a04718/Common/Core/vtkLookupTable.cxx#L707-L732

Given a vtkDataArray with the lookup table colors, set the internal
table based on these colors. The passed array should have unsigned char
type and 4 components per tuple for the RGBA colors.

The internal table is then augmented with special colors. The internally
table is kept as a `Array` versus a `vtkDataArray` as in the C++
implementation. This could possibly be a target for future refactoring.

ScalarToColors `setAnnotation` is fixed -- this was never covered and
had conflicting internal assumptions that its arguments were
vtkDataArray's or (Array or TypedArray's). Now, Array or TypedArray are
supported.