feat(TextureLODsDownloader): Add a downloader for low-to-high resolution textures
The TextureLODsDownloader is designed to download a series of texture images from low to high
resolution, and update the supplied vtkTexture as each download finishes. It currently downloads the
texture images sequentially rather than in parallel. There is also an option to provide a callback
for when each step finishes.

This feature was added to the HttpSceneLoader.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>