feat(InteractorStyleManipulator): Add getters for manipulators
This allows manipulators to be accessed via getters. This is needed particularly because the
KeyboardCameraManipulator has a movement speed associated with it, and this movement speed may
sometimes need to be set by users. If the KeyboardCameraManipulator was generated using a preset, a
function like getKeyboardManipulators() may be the only way to access the manipulator to set the
movement speed.

Getters were added for other manipulators as well in case they are also needed in the future.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>