feat(KeyboardCameraManipulator): Add ability to reset movement speed
resetMovementSpeed() generates a movement speed based upon the longest length of the rendering
bounding box. It is intended to just be a reasonable guess as to what a good movement speed would
be. The user is able to modify the movement speed via setMovementSpeed(). If the movement speed is
not set in the options when the KeyboardCameraManipulator is constructed, upon the first movement,
resetMovementSpeed() will be used to generate a reasonable speed.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>