feat(LODsLoaders): Add wait time options to start and between downloads
This is being added primarily so that a wait time to start can be set. It seems much nicer if the
LODs wait to start downloading/updating until the scene and renderer are all set up. It might be
good to have a function call start the downloads, but for now, a timer seems to be working fine.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>