perf(TextureLODsDownloader): Decode images before setting on texture
The HTMLImageElement decode function appears to have been written exactly for our need: to try to
prevent freezing when the browser tries to render an encoded image. The function decodes the image
asynchronously, and then we just use the decoded image. This function is new, but it appears that
all major browsers currently support it. However, my performance tests indicate that a decode is
still happening at render time and freezing the rendering. I am not entirely sure why.