fix(VolumeMapper): Emphasize fragments based on labelmap pwf
For any "proportional" components, this change applies a scaling factor
to the opacity of fragments based on both the piecewise function value
and the component weight.  When the weight of the proportional component
is small, the effect on fragment opacity from the component piecewise
function is small, and conversely, when the weight of the proportional
component weight is large, the effect is more pronounced.