fix(RenderWindowInteractor): Reset lastFrameStart with tab switch
When tabs are not active browser will step calling
requestAnimationFrame callbacks. See:

  https://developer.mozilla.org/en-US/docs/Web/API/Window/requestAnimationFrame

This causes the RenderWindowInteractor lastFrameStart to be inaccurate,
which results in an incorrect lastFrameTime. This can cause the
VolumeMapper, which adjusts its the sampling distance based on the
framerate, to assume too low of a frame rate and produce poor quality
renderings.

Use the Page Visibility API:

  https://developer.mozilla.org/en-US/docs/Web/API/Page_Visibility_API

to detected when tabs are switched, and reset the RenderWindowInteractor
lastFrameStart.

Co-authored-by: Ken Martin <ken.martin@kitware.com>